Import of MLEHMANN/Audio-Play-MPG123-0.052 from CPAN.
gitpan-cpan-distribution: Audio-Play-MPG123
gitpan-cpan-version:      0.052
gitpan-cpan-path:         MLEHMANN/Audio-Play-MPG123-0.052.tar.gz
gitpan-cpan-author:       MLEHMANN
gitpan-cpan-maturity:     released

+++ mpg123/term.c
@@ -3,9 +3,13 @@
 
 #include <termios.h>
 #include <unistd.h>
+#include <stdlib.h>
 #include <ctype.h>
 #include <sys/time.h>
 #include <sys/types.h>
+#include <sys/wait.h>
+
+
 
 #include "mpg123.h"
 #include "buffer.h"
@@ -14,6 +18,7 @@
 
 static int term_enable = 0;
 static struct termios old_tio;
+static void term_quit(void);
 
 /* initialze terminal */
 void term_init(void)
@@ -22,6 +27,7 @@ void term_init(void)
 
   term_enable = 0;
 
+  atexit(term_restore);
   if(tcgetattr(0,&tio) < 0) {
     fprintf(stderr,"Can't get terminal attributes\n");
     return;
@@ -129,8 +135,11 @@ static long term_handle_input(struct frame *fr, int do_delay)
 	  kill(0,SIGINT);
 	  break;
 	case QUIT_KEY:
-	  kill(0,SIGTERM);
-	  break;
+	  if (val!=QUIT_KEY)	/* Uppercase stops dead. */
+	  	kill(0,SIGTERM);
+	  else 
+          	term_quit();
+          exit(1);
 	case PAUSE_KEY:
   	  paused=1-paused;
 	  if(paused) {
@@ -183,10 +192,33 @@ static long term_handle_input(struct frame *fr, int do_delay)
 void term_restore(void)
 {
   
-  if(!term_enable)
+  if (!term_enable || !param.term_ctrl)
     return;
 
+  term_enable = 0;
   tcsetattr(0,TCSAFLUSH,&old_tio);
+  if (param.verbose)
+    puts("");
+}
+
+static void term_quit() {
+    extern int buffer_pid;
+    extern struct audio_info_struct ai;
+    /* quit gracefully */
+    if (param.usebuffer) {
+        buffer_end();
+        xfermem_done_writer(buffermem);
+        waitpid(buffer_pid, NULL, 0);
+        xfermem_done(buffermem);
+    } else {
+        audio_flush(param.outmode, &ai);
+    }
+    if (param.outmode == DECODE_AUDIO)
+        audio_close(&ai);
+    if (param.outmode == DECODE_WAV)
+        wav_close();
+    term_restore();
+    exit(0);
 }
 
 #endif
